[~] move DTO to root package
