+++ effectiveOfficeBackend/src/main/kotlin/office/effective/features/booking/facade/BookingFacade.kt
@@ -4,7 +4,7 @@ import office.effective.common.exception.InstanceNotFoundException
 import office.effective.common.utils.DatabaseTransactionManager
 import office.effective.common.utils.UuidValidator
 import office.effective.features.booking.converters.BookingFacadeConverter
-import office.effective.features.booking.dto.BookingDTO
+import office.effective.dto.BookingDTO
 import office.effective.features.booking.service.BookingService
 import office.effective.model.Booking
 import office.effective.model.Workspace
