+++ effectiveOfficeBackend/src/main/kotlin/office/effective/features/booking/routes/BookingRouting.kt
@@ -10,7 +10,7 @@ import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import office.effective.common.swagger.SwaggerDocument
-import office.effective.features.booking.dto.BookingDTO
+import office.effective.dto.BookingDTO
 import office.effective.features.booking.facade.BookingFacade
 import office.effective.features.booking.routes.swagger.*
 import org.koin.core.context.GlobalContext
