+++ effectiveOfficeBackend/src/main/kotlin/office/effective/features/user/facade/UserFacade.kt
@@ -3,7 +3,7 @@ package office.effective.features.user.facade
 import office.effective.common.utils.DatabaseTransactionManager
 import office.effective.features.user.ITokenVerifier
 import office.effective.features.user.converters.UserDTOModelConverter
-import office.effective.features.user.dto.UserDTO
+import office.effective.dto.UserDTO
 import office.effective.features.user.service.IUserService
 import office.effective.model.UserModel
