+++ effectiveOfficeBackend/src/main/kotlin/office/effective/features/user/routes/swagger/UserSwagger.kt
@@ -3,9 +3,8 @@ package office.effective.features.user.routes.swagger
 import io.github.smiley4.ktorswaggerui.dsl.OpenApiRoute
 import io.ktor.http.*
 import office.effective.common.swagger.SwaggerDocument
-import office.effective.features.booking.dto.BookingDTO
-import office.effective.features.user.dto.IntegrationDTO
-import office.effective.features.user.dto.UserDTO
+import office.effective.dto.IntegrationDTO
+import office.effective.dto.UserDTO
 
 fun SwaggerDocument.returnUserByEmail(): OpenApiRoute.() -> Unit = {
     description = "Return user by email"
